`ct/l1`: abstract out `*_builder` utilities for `metastore` tests
This will be useful in a new test file added in a future commit (and
potentially for future test suites).
